// example/src/v2/pages/demo-page/types/color-path.ts
export type ColorPath =
  | 'primary.default'
  | 'primary.light'
  | 'primary.dark'
  | 'primary.contrastText'
  | 'secondary.default'
  | 'secondary.light'
  | 'secondary.dark'
  | 'secondary.contrastText'
  | 'grey.100'
  | 'grey.200'
  | 'grey.300'
  | 'grey.400'
  | 'grey.500'
  | 'grey.600'
  | 'grey.700'
  | 'grey.800'
  | 'grey.900'
  | 'info.primary'
  | 'info.light'
  | 'success.primary'
  | 'success.light'
  | 'warning.primary'
  | 'warning.light'
  | 'error.primary'
  | 'error.light'
  | 'white';